requirements.txt
blessed
pyserial
git+https://github.com/LiamBindle/PyVESC
